requirements.txt
selene == 2.0.0rc9
allure-pytest
pytest~=8.3.5
allure-python-commons~=2.14.0
selenium~=4.31.0